requirements.txt
flask
flask-restful
flask-migrate
flask-sqlalchemy
flask-marshmallow
flask-jwt-extended
Flask-Cors
marshmallow-sqlalchemy
marshmallow-jsonapi
python-dotenv
passlib
tox
